Move game matmul tests
